feat(cart): count in kilos and metres, not only in pieces
Quantities are decimals across the app now. A card steps in the product's own
unit - a piece, a quarter kilo, 50 g - and prints what it counts: "1.5 كيلو".
The cart, the product screen and the order lines all read the same way, and
the wire carries the fraction through to the store.

+++ lib/app/extensions/cart_branch_guard.dart
@@ -8,7 +8,7 @@ Future<void> addToCartGuarded(
   BuildContext context,
   WidgetRef ref,
   BranchProduct product,
-  int quantity,
+  double quantity,
 ) async {
   final cart = ref.read(cartController);
   final cartBranchId = cart.cartBranchId;
